1st commit of v1.1:  - coupon support  - move prices from menu item -> location item  - refactor menu & order code to use location item information  - make heating instructions dynamic

+++ RyteBytes/LoginViewController.m
@@ -20,7 +20,7 @@
 
 NSString *loginSuccessSegue = @"LoginSuccess";
 
-int tagTextFieldToResign;
+long tagTextFieldToResign;
 
 - (id)initWithNibName:(NSString *)nibNameOrNil bundle:(NSBundle *)nibBundleOrNil
 {
